fast_api_postgre_login/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.91.0
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.4
sniffio==1.3.0
starlette==0.24.0
typing-extensions==4.4.0
uvicorn==0.20.0
